+++ admin/tabs/activity/clienterrors/ClientErrorsModel.js
@@ -188,7 +188,7 @@ export class ClientErrorsModel extends HoistModel {
             });
 
             gridModel.loadData(data);
-            await gridModel.selectFirstAsync({ifNoSelection: true});
+            await gridModel.preSelectFirstAsync();
 
         } catch (e) {
             gridModel.clear();